Refactor API routes to simplify auth checks
A single catch-all API route worked nicely until we started needing
additional checks for permissions and the injection of environment
variables in certain cases. This commit extracts a higher order
function which wraps API functions. It performs high level authorisation
and wraps the API function in a generic try-catch to deal with any
service API errors.

The idea is that we can work with a smaller function body for
each API route as we fine-tune each endpoint to cater
for different user permissions.

The API route files should then mostly consist of checking user
permissions for certain requests, and injecting any necessary params
or perhaps headers for the service API request.