Add a page authorisation system
Allows per-page authorisation configuration by adding a
'permittedRoles' property, which is now required in all *page*
components, otherwise logged in users will not be able to access them.

Note that are two exceptions to the above rule: the login page, and
the access-denied page component, (see GoogleAuth.js).

The 'permittedRoles' lists out the user roles that have access to the
page. We will use filtering in the API layer to control exactly which
resources a user can access on a page they have access to.

This commit does NOT specify the final permissions for our pages
because those are not currently known. Instead, it allows access for
all users, and in that sense this is a "no op".

Also:
  - Configure jest to mock the style fetched in _app.js instead of
    referencing 'identity-obj-proxy' which I don't think we're using [1]
  - Show logged in users the access-denied page if they try to access
    something they do not have permissions to view.
  - Do not show logged out users an access-denied page (always
    redirect to the login page)
  - Remove a sentence from access denied page which no longer applies

[1] https://jestjs.io/docs/en/webpack#handling-static-assets