Prevent occasional client side transition errors
Previously if a user tried to navigate to a page and they fail an auth
check (e.g. they cleared their cookies), the entire request fails. A
developer working locally would see the whole app fail with
little clues as to why.

The issue seems to happen because we would write a redirect, but then
Next JS would write another response on top. Writing to the response
multiple times is not allowed in Node, and a server error is raised.

This commit aims to fix this by testing whether the request came from
a client side transition. If it does, then an unauthorised user
should see the access denied message, rather than getting redirected
there.

To enable that a tiny "access denied" component has been extracted.