Detect contractor groups using a partial match
We've started to go down the route of adding individual environment
variables per contractor, but in fact the frontend shouldn't need to
know which particular contractor the user belongs to for its
current functionality.

This commit replaces our specific checks for contractor groups with
a partial match on the group name prefix. It relies on a consistent
group naming scheme and a new environment variable to hold that prefix.

The intention is to remove the need to manage lots of environment
variables as contractors are added and removed.

The backend is relied upon to perform specific group checks instead.

Includes:
  - Removal of specific contractor group names from .env.sample