Fetch work order when closing
Otherwise it's possible to change the work order ID in the url and
show the close page for a non-existent work order. This surfaces
the issue earlier by showing an error message and will support
later work which requires additional props from the work order at the
point of closure.