Avoid routing to a separate print page
Render a print-only version of the work order from the
show page and trigger an image load for the second page only when
needed (i.e. when the user hits 'print' from the work order actions
menu).

This saves us requesting the same information about the work order and
also saves the user a couple of clicks.

Includes:
  - Supplying of tasks and sors to the tabs component, because this
  now needs to be fetched by the parent to populate the printed version.
  Since that's the default tab anyway, this doesn't seem like too
  strange a thing to do.