Avoid managing two priority form fields
Previously the priority select part of the form had both its value
and text value set to the priority description we get from the
API for each priority.

This required some logic and a hidden field to track both the selected
description and its code. By levereging the ability of the select to
hold a text value different from its internal value (description
and code respectively), we can simplify some of this form and its state.

Also:
 - Rename priority field to priorityCode. Ultimately the user is
 selecting a code here so this probably is more readable this way.