Avoid completion calculations using priority maps
Split apart parameters supplied to the completion calculation function
so that there's no longer a need to use the mapping between
priority codes and numbers of hours / days. The calling code will
be able to use the API object to supply these, meaning we no longer
need to maintain a code-time mapping for priorities on the frontend.