Invert logic for detecting low priority
We want to get away from storing priority information on the
frontend, especially as more priorities are being added to the API. But
for the purposes of detecting if an order is affected by low
priority holidays we still need some kind of categorisation of
priorities. This commit achieves that by making a list of the high
priorities. This should require less maintenance as we expect those to
be more stable than a list of low priorities.

A better longer term solution would be to receive the
priority type from the backend, or even for it to calculate its own
completion times, but we're not there currently.