Swap workspace persistence for checkout & node module cache
The workspace takes over a minute to persist and around 10 seconds to restore. This is significantly slower than the 2s checkout time. Caching is reported to be quicker than workspace persistence, so this chance reduces the scope of the cache and updates it to match the recommendations in the official documentation