🏃‍♂️ Speed up CircleCI (even more) (#961)
* Swap workspace persistence for checkout & node module cache

The workspace takes over a minute to persist and around 10 seconds to restore. This is significantly slower than the 2s checkout time. Caching is reported to be quicker than workspace persistence, so this chance reduces the scope of the cache and updates it to match the recommendations in the official documentation

* Prefix node commands with yarn for CI consistency

This reduces the chance of running a developer-specific version and ensures all dependencies are managed with Yarn

* Remove dependency install step

Installing dependencies once before splitting out to multiple jobs is intuitively faster, however it actually introduces an extra overhead in both the best and worst case scenarios:

- Best case: the cache exists and no extra packages are installed. The install-dependencies step takes around 15 seconds to spin up, restore the cache, and assert that the install is a no-op. This means adding around 15 seconds to most builds, as dependencies are changed fairly rarely.
- Worst case: the cache doesn't exist and all packages must be instsalled. This should be the greatest performance improvement, with the install-dependencies step taking around 1:30 to spin up, install and save the cache. Subsequent steps then take 15 seconds to load the cache. However, without this step both jobs take 1:30 to run and update the cache, without the additional 15 second overhead. As the jobs run in parallel the overall execution time isn't affected.

There is a downside to this approach with parallel tests: when the cache is invalidated we'll use up an extra mi
nute of compute for each parallel worker (and each parallel job). Therefore this may increase overall costs - this will need to be balanced against the reduced compute in all other runs

* Only install Chrome/Chromedriver in CircleCI

Firefox adds an extra 20 seconds to every e2e job and we're not currently testing against it. This change adds only installs the Chrome-relevant components, for speed 🏃.

* Remove Node downgrade for deploy

This looks like an old attempt to manage node versions in CI. Now we use official CircleCI node images, so we don't need to manually install NVM. Also this downgrades node to an unsupported old version, so removing this step keeps tests and deploy using a consistent Node version.

* Fix JUnit output for CircleCI test splitter

The default JUNit output for Cypress doesn't include the file path for each test suite. This results in CircleCI
being unable to associate test results and times to the files they come from. We need this for CircleCI's time-based
test splitting, which balances Cypress tests across nodes to minimise the test run time.

This change introduces a workaround post-processing script from https://github.com/michaelleeallen/mocha-junit-reporter/issues/132\#issuecomment-721943600.