Remember data when photo upload fails (#1126)
HPT-763 Allow the app to remember the data even if the photo upload fails

Uses the react-hook-form-persist package to back-up the mobile work order form to browser local storage, so that if there's an error with the file upload, the form can re-set to the previous values set by the user.

This change is to mitigate time lost by operatives after photo upload fails (e.g. due to internet disconnection) and the resulting frustration, which should also reduce the risk of mandatory photos.

As a bonus, the data is also preserved if the operative exits the app and returns, refreshes the page, etc, however in this case the photos are not preserved because they aren't supported in react-hook-form-persist and are preserved using React state instead.